# public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FutureStarr API",
        "description": "FutureStar API Swagger Documentation. IF you want to use postman just import documetation json file",
        "contact": {
            "email": "admin@admin.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/buyer-account/details": {
            "get": {
                "tags": [
                    "Buyer Module"
                ],
                "summary": "BuyerAccount",
                "description": "BuyerAccount",
                "operationId": "Buyer Account",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "description": "Authorization token obtained from logging in.",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}